Handle open wiki links
Previously, open wiki links caused an unhandled exception due to not closing tokens. We now
pass through to other tokenizers when there is a newline or we're at the end of the file.
This commit also adjusts some of the `consume` logic to satisfy the assertions in micromark
when running tests with 'micromark/lib'.